pyproject.toml
[tool.poetry]
name = "dbt-jobs-as-code"
version = "0.1.0"
description = "A CLI to allow defining dbt Cloud jobs as code"
authors = []
license = "Apache License 2.0"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.7"
click = "^8.1.3"
requests = "^2.28.1"
loguru = "^0.6.0"
deepdiff = "^6.2.2"
pydantic = "^1.10.2"
croniter = "^1.3.8"
"ruamel.yaml" = "^0.17.21"
rich = "^12.6.0"
PyYAML = "^6.0"
jsonschema = "^4.17.3"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pytest = "^7.2.0"

[tool.poetry.scripts]
dbt-jobs-as-code = "src.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]

[tool.black]
line-length = 99
